RequestError class
