+++ lib/ya/api/direct.rb
@@ -1,3 +1,3 @@
 require "ya/api/direct/version"
 require "ya/api/direct/client"
-require "ya/api/direct/exception"
+require "ya/api/direct/exceptions"
